+++ src/main/java/com/informatorio/trabajofinal/controller/SourceController.java
@@ -5,11 +5,12 @@ import com.informatorio.trabajofinal.domain.Source;
 import com.informatorio.trabajofinal.dto.SourceDTO;
 import com.informatorio.trabajofinal.repository.SourceRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.List;
-import java.util.Locale;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @RestController
@@ -25,43 +26,47 @@ public class SourceController {
     }
 
     @PostMapping("/source")
-    public SourceDTO createSource(@RequestBody Source source) {
-        source = new Source(source.getName(), source.getCreatedAt());
+    public ResponseEntity<SourceDTO> createSource(@RequestBody @Valid SourceDTO sourceDTO) {
+        Source source = sourceConverter.toEntity(sourceDTO);
         sourceRepository.save(source);
-        return sourceConverter.toDto(source);
+        return new ResponseEntity<>(sourceConverter.toDto(source), HttpStatus.CREATED);
     }
 
     @DeleteMapping("/source/{id}")
-    public SourceDTO deleteSource(@PathVariable Integer id) {
+    public ResponseEntity<?> deleteSource(@PathVariable Integer id) {
         Source capture = sourceRepository.findById(id).orElse(null);
-        sourceRepository.deleteById(id);
-        return sourceConverter.toDto(capture);
+        if (capture != null) {
+            sourceRepository.deleteById(id);
+            return new ResponseEntity<>(sourceConverter.toDto(capture), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("Source inexistente", HttpStatus.NOT_FOUND);
+        }
     }
 
     @PutMapping("/source/{id}")
-    public SourceDTO updateSource(@PathVariable Integer id, @RequestBody Source newSource) {
+    public ResponseEntity<?> updateSource(@PathVariable Integer id, @RequestBody @Valid SourceDTO sourceDTO) {
         Source source = sourceRepository.findById(id).orElse(null);
         if (source != null) {
-            source.setName(newSource.getName());
-            source.setCode(newSource.getName());
-            source.setCreatedAt(newSource.getCreatedAt());
+            source.setName(sourceDTO.getName());
+            source.setCode(sourceDTO.getName());
+            source.setCreatedAt(sourceDTO.getCreatedAt());
+            sourceRepository.save(source);
+            return new ResponseEntity<>(sourceConverter.toDto(source), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("Source no encontrado", HttpStatus.NOT_FOUND);
         }
-        sourceRepository.save(source);
-        return sourceConverter.toDto(source);
     }
 
     @GetMapping("/source")
-    public List<SourceDTO> getAllSources() {
-        return sourceConverter.listSourcesToDto(sourceRepository.findAll());
-
-
+    public ResponseEntity<List<SourceDTO>> getAllSources() {
+        return new ResponseEntity<>(sourceConverter.listSourcesToDto(sourceRepository.findAll()), HttpStatus.OK);
     }
 
-    @GetMapping("/source/str:{string}")
-    public List<SourceDTO> getAllSourcesWithStringInName(@PathVariable CharSequence string) {
+    @GetMapping(value = "/source", params = {"string"})
+    public ResponseEntity<List<SourceDTO>> getAllSourcesContainsStringInName(@RequestParam String string) {
         List<Source> sources = sourceRepository.findAll().stream()
-                .filter(source -> source.getCode().contains(string))
+                .filter(source -> source.getCode().contains(string.toLowerCase()))
                 .collect(Collectors.toList());
-        return sourceConverter.listSourcesToDto(sources);
+        return new ResponseEntity<>(sourceConverter.listSourcesToDto(sources), HttpStatus.OK);
     }
 }
